requirements.txt
python-dotenv
pandas
numpy
sqlalchemy
psycopg2-binary
